Get search  arguments from command line.
Get search terms from command line (using -a flag) and create the
appropriate request to the API.
The argument passed to the command line must be one of the following:
"ada", "org", "subject", "protocol", "term", "unit", "signer","type",
"tag", "from_date", "to_date", "from_issue_date", "status", "size".
If an argument is passed to the spider and does not exist to the
aforementioned list, a warning is issued on the log.
If no arguments are supplied to the spider, the spider requests the
decisions of the previous day.